[#510] Separate privacy optionsbfrom form component
